Move static pages into static folder and some file name cleanup
