--- src/pages/home/HomePage.spec.tsx
@@ -1,21 +1,0 @@
-import * as HomePage from './HomePage';
-import React from 'react';
-import { assert } from 'chai';
-import { shallow, mount } from 'enzyme';
-import sinon from 'sinon';
-
-describe('HomePage', () => {
-
-    before(()=>{
-
-    });
-
-    after(()=>{
-
-    });
-
-    it('what does it do?', ()=>{
-
-    });
-
-});
