+++ src/pages/home/HomePage.tsx
@@ -1,63 +1,92 @@
 import * as React from 'react';
-import exposeComponentRenderer from 'shared/lib/exposeComponentRenderer';
-import {FlexCol, FlexRow} from "../../shared/components/flexbox/FlexBox";
-import {observer, inject} from "mobx-react";
-import DevTools from "mobx-react-devtools";
-import {toJS, observable, action, computed, whyRun, expr} from "mobx";
-import LabeledCheckbox from "../../shared/components/labeledCheckbox/LabeledCheckbox";
-import ReactSelect from 'react-select';
+import * as _ from 'lodash';
+import {If, Then, Else} from 'react-if';
+import {observer, inject, Observer } from "mobx-react";
+import { observable } from 'mobx';
+import Chart from 'chart.js';
+import AppConfig from 'appConfig';
 import 'react-select/dist/react-select.css';
+import {CancerStudyQueryUrlParams, QueryStore} from "../../shared/components/query/QueryStore";
 import QueryAndDownloadTabs from "../../shared/components/query/QueryAndDownloadTabs";
-import {QueryStore} from "../../shared/components/query/QueryStore";
-import QueryModal from "../../shared/components/query/QueryModal";
-import BarGraph from "shared/components/barGraph/BarGraph";
-import client from '../../shared/api/cbioportalClientInstance';
-import {remoteData} from "../../shared/api/remoteData";
+import {PageLayout} from "../../shared/components/PageLayout/PageLayout";
 import RightBar from "../../shared/components/rightbar/RightBar";
-import AppConfig from "appConfig";
+import getBrowserWindow from "../../shared/lib/getBrowserWindow";
 
+(Chart as any).plugins.register({
+    beforeDraw: function (chartInstance: any) {
+        const ctx = chartInstance.chart.ctx;
+        ctx.fillStyle = "white";
+        ctx.fillRect(0, 0, chartInstance.chart.width, chartInstance.chart.height);
+    }
+});
 
-export class HomePageStore {
+const win = (window as any);
 
-    readonly data = remoteData({
-        invoke: () => {
-            return client.getAllStudiesUsingGET({projection: "DETAILED"});
-
-        }
-    });
+export interface IResultsViewPageProps {
+    routing: any;
 }
 
-interface IHomePageProps
-{
-    queryStore:QueryStore
+export function createQueryStore() {
+
+    const win:any = window;
+
+    // lets make query Store since it's used in a lot of places
+    const queryStore = new QueryStore(win, getBrowserWindow().routingStore.query);
+
+    queryStore.singlePageAppSubmitRoutine = function(path:string, query:CancerStudyQueryUrlParams) {
+
+        // normalize this
+        query.cancer_study_list = query.cancer_study_list || query.cancer_study_id;
+        delete query.cancer_study_id;
+
+        // need to come up with better place to put app globals (things used by different routes)
+        win.currentQueryStore = queryStore;
+
+        win.routingStore.updateRoute(query, "results");
+
+    };
+
+    return queryStore;
+
 }
 
-interface IHomePageState
-{
-}
+@inject('routing')
+@observer
+export default class HomePage extends React.Component<IResultsViewPageProps, {}> {
 
-@inject("queryStore") @observer
-export default class HomePage extends React.Component<IHomePageProps, IHomePageState>
-{
-    constructor(props:IHomePageProps)
-    {
+    private showTwitter = AppConfig.showTwitter === true;
+
+    @observable showQuerySelector = true;
+
+    @observable currentQuery:boolean;
+
+    queryStore:QueryStore;
+
+    constructor(props: IResultsViewPageProps) {
         super(props);
     }
 
-    getModalWrappedComponent(){
-        return (
-            <QueryModal store={this.props.queryStore} />
-        )
+    componentWillMount(){
+        this.queryStore = createQueryStore();
     }
 
-    public render()
-    {
-        const blurb:JSX.Element|null = AppConfig.skinBlurb? <p style={{marginBottom:"20px"}} dangerouslySetInnerHTML={{__html: AppConfig.skinBlurb}}></p> : null;
-        return (<div>
-                   <div>
-                       {blurb}
-                   </div>
-                   <QueryAndDownloadTabs store={this.props.queryStore} />
-                </div>);
+    private handleTabChange(id: string) {
+        this.props.routing.updateRoute({ tab: id });
+    }
+
+    public render() {
+
+        return (
+            <PageLayout noMargin={true} rightBar={<RightBar queryStore={this.queryStore} />}>
+                <div style={{padding:"0 15px"}}>
+                    <p style={{marginBottom:15}}>The cBioPortal for Cancer Genomics provides <b>visualization</b>, <b>analysis</b> and <b>download</b> of large-scale <b>cancer genomics</b> data sets.
+                        <br />
+                        <b>Please cite</b> <a href="http://www.ncbi.nlm.nih.gov/pubmed/23550210">Gao et al. <i>Sci. Signal.</i> 2013</a> &amp;  <a href="http://cancerdiscovery.aacrjournals.org/content/2/5/401.abstract"> Cerami et al. <i>Cancer Discov.</i> 2012</a> when publishing results based on cBioPortal.</p>
+                    <QueryAndDownloadTabs store={this.queryStore}/>
+                </div>
+                <hr />
+            </PageLayout>
+        )
+
     }
 }
