+++ src/pages/staticPages/aboutus/AboutUs.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
-import {PageLayout} from "../../shared/components/PageLayout/PageLayout";
+import {PageLayout} from "../../../shared/components/PageLayout/PageLayout";
 import './styles.scss';
 import AppConfig from "appConfig";
-import StaticContent from "../../shared/components/staticContent/StaticContent";
+import StaticContent from "../../../shared/components/staticContent/StaticContent";
 import Helmet from "react-helmet";
 
 export default class AboutUs extends React.Component<{}, {}> {
