+++ src/pages/staticPages/datasetView/styles.module.scss
