+++ src/pages/staticPages/news/styles.scss
