+++ src/pages/staticPages/rmatlab/RMatLAB.tsx
@@ -1,12 +1,12 @@
 import * as React from 'react';
 import {observer} from 'mobx-react';
-import {PageLayout} from "../../shared/components/PageLayout/PageLayout";
+import {PageLayout} from "../../../shared/components/PageLayout/PageLayout";
 import './styles.scss';
 import Helmet from "react-helmet";
 
 @observer
 export default class CGDS extends React.Component<{}, {}> {
-
+//
     public render() {
 
         return <PageLayout className={'whiteBackground'}>
