+++ src/pages/staticPages/tools/mutationMapper/resources/standaloneMutationDataExample.txt
