+++ src/pages/staticPages/tools/mutationMapper/resources/standaloneMutationDataExampleWithGeneAndProteinChangeOnly.txt
