+++ src/pages/staticPages/tools/mutationMapper/resources/standaloneMutationDataExampleWithGenomicCoordinatesOnly.txt
