+++ src/pages/staticPages/tutorials/styles.scss
