+++ src/pages/staticPages/visualize/styles.scss
