+++ src/pages/staticPages/webAPI/styles.scss
